Issue #218: Moved methods with Git commands to separate GitCommands class
